[rocm-libraries] ROCm/rocm-libraries#994 (commit e0d1494)
Enable pencil-to-pencil global transposition with
 sub-communicators (#994)
MIME-Version: 1.0
Content-Type: text/plain; charset=UTF-8
Content-Transfer-Encoding: 8bit

Summary of proposed changes:

- The objective of this PR is to add support for All-to-All
communication with sub-communicators for uniform Pencil decompositions,
as shown in the following figure.

<img width=1791 height=1013 alt=image
src=https://github.com/user-attachments/assets/3a01414c-bdac-4b2f-80df-1cdd1a1c2608
/>

The proposed updates work as follows:

1. We have added a function that reads (or finds) the input and output
grids for and . It then returns a transposition plan for pencil
decomposition, including brick-to-pencil, pencil-to-pencil,
slab-to-pencil, etc. For example: if and , it returns the following
sequence of intermediate grids: along with a corresponding sequence of
transpose types. The selection of intermediate grids is based on a
heuristic approach.

2. In the case of uniform pencils; i.e., when there is a uniform volume
of messages to send and a subset of neighbors to send to, we create
pencil sub-communicators and perform a global transpose using .

3. This PR focuses only on pencil-to-pencil transposition. Therefore, we
have left the default path intact, which relies on either or when the
global communicator allows for uniform exchanges (slab case).

4. Implementing an optimized pencil-to-pencil with MPI_Alltoall is not
just for performance gains, but also to ensure scalability.